requirements.txt
timm
torchvision
pillow
medmnist
opendatasets
icrawler
matplotlib
seaborn
pymupdf
selenium
aiohttp
streamlit
psutil
